Refactor use case methods for consistency and clarity
- Renamed methods in ManageArtifactsUseCase for better clarity:
  - create() to createArtifact()
  - getById() to getArtifact()
  - list() to listArtifacts()
  - listByType() to listArtifacts()
  - update() to updateArtifact()

- Updated method names in ManageAutomationsUseCase:
  - create() to createAutomation()
  - getById() to getAutomation()
  - list() to listAutomations()
  - update() to updateAutomation()

- Adjusted method names in ManageDecisionsUseCase:
  - create() to createDecision()
  - getById() to getDecision()
  - list() to listDecisions()
  - update() to updateDecision()

- Modified method names in ManageFormsUseCase:
  - create() to createForm()
  - getById() to getForm()
  - list() to listForms()
  - update() to updateForm()

- Changed method names in ManageProcessesUseCase:
  - create() to createProcess()
  - getById() to getProcess()
  - list() to listProcesses()
  - listByProject() to listProcesses()
  - update() to updateProcess()
  - deploy() to deployProcess()

- Updated method names in ManageTasksUseCase:
  - create() to createTask()
  - getById() to getTask()
  - list() to listTasks()
  - listByAssignee() to listTasksByAssignee()
  - listByProcessInstance() to listTasksByProcessInstance()
  - claim() to claimTask()
  - complete() to completeTask()
  - update() to updateTask()

- Refined method names in ManageTriggersUseCase:
  - create() to createTrigger()
  - getById() to getTrigger()
  - list() to listTriggers()
  - listByProcess() to listTriggers()
  - update() to updateTrigger()

- Adjusted method names in ManageVisibilitiesUseCase:
  - create() to createVisibility()
  - getById() to getVisibility()
  - list() to listVisibilities()
  - update() to updateVisibility()

- Updated method names in ManageAutomationRulesUseCase:
  - create() to createAutomationRule()
  - getById() to getAutomationRule()
  - list() to listAutomationRules()
  - listByTemplate() to listAutomationRules()
  - listActive() to listActiveAutomationRules()
  - update() to updateAutomationRule()

- Refined method names in ManageDashboardsUseCase:
  - create() to createDashboard()
  - getById() to getDashboard()
  - list() to listDashboards()
  - update() to updateDashboard()

- Adjusted method names in ManageDataContextsUseCase:
  - create() to createDataContext()
  - getById() to getDataContextById()
  - list() to listDataContexts()
  - listByInstance() to listDataContextsByInstance()

- Updated method names in ManageEntityTypesUseCase:
  - create() to createEntityType()
  - getById() to getEntityTypeById()
  - list() to listEntityTypes()
  - update() to updateEntityType()

- Refined method names in ManageNotificationsUseCase:
  - create() to createNotification()
  - getById() to getNotificationById()
  - list() to listNotifications()
  - listByRecipient() to listNotifications()
  - listByInstance() to listNotifications()
  - update() to updateNotification()

- Adjusted method names in ManageSituationActionsUseCase:
  - create() to createSituationAction()
  - getById() to getSituationActiond()
  - list() to listSituationActions()
  - update() to updateSituationAction()

- Updated method names in ManageSituationInstancesUseCase:
  - create() to createSituationInstance()
  - getById() to getSituationInstance()
  - list() to listSituationInstances()
  - listByTemplate() to listSituationInstances()
  - listByStatus() to listSituationInstances()
  - update() to updateSituationInstance()
  - resolve() to resolveSituationInstance()

- Refined method names in ManageSituationTemplatesUseCase:
  - create() to createSituationTemplate()
  - getById() to getSituationTemplate()
  - list() to listSituationTemplates()
  - listByEntityType() to listSituationTemplates()
  - update() to updateSituationTemplate()

- Updated method names in ManageSubstitutionRulesUseCase:
  - getById() to getSubstitutionRuleById()
  - listByUser() to listSubstitutionRulesByUser()
  - listBySubstitute() to listSubstitutionRulesBySubstitute()
  - listByStatus() to listSubstitutionRulesByStatus()
  - create() to createSubstitutionRule()
  - update() to updateSubstitutionRule()
  - activate() to activateSubstitutionRule()
  - deactivate() to deactivateSubstitutionRule()

- Adjusted method names in ManageTaskAttachmentsUseCase:
  - getById() to getTaskAttachmentById()
  - listByTask() to listTaskAttachmentsByTask()
  - create() to createTaskAttachment()

- Updated method names in ManageTaskCommentsUseCase:
  - getById() to getTaskCommentById()
  - listByTask() to listTaskCommentsByTask()
  - create() to createTaskComment()
  - update() to updateTaskComment()

- Refined method names in ManageTaskDefinitionsUseCase:
  - getById() to getTaskDefinitionById()
  - list() to listTaskDefinitions()
  - listByProvider() to listTaskDefinitionsByProvider()
  - listByCategory() to listTaskDefinitionsByCategory()
  - create() to createTaskDefinition()
  - update() to updateTaskDefinition()

- Updated method names in ManageTaskProvidersUseCase:
  - create() to createTaskProvider()
  - update() to updateTaskProvider()

- Adjusted method names in ManageTasksUseCase:
  - getById() to getTaskById()
  - list() to listTasks()
  - listByAssignee() to listTasksByAssignee()
  - listByStatus() to listTasksByStatus()
  - listByProvider() to listTasksByProvider()
  - listByCategory() to listTasksByCategory()
  - listByPriority() to listTasksByPriority()
  - create() to createTask()
  - update() to updateTask()
  - release() to releaseTask()
  - forward() to forwardTask()
  - complete() to completeTask()
  - cancel() to cancelTask()

- Updated method names in ManageUserTaskFiltersUseCase:
  - getById() to getUserTaskFilterById()
  - listByUser() to listUserTaskFiltersByUser()
  - getDefault() to getDefaultUserTaskFilter()
  - create() to createUserTaskFilter()
  - setDefault() to setDefaultUserTaskFilter()

+++ mobile/source/uim/platform/mobile/application/usecases/manage/mobile_apps.d
@@ -19,7 +19,7 @@ class ManageMobileAppsUseCase { // TODO: UIMUseCase {
         this.repo = repo;
     }
 
-    CommandResult create(CreateMobileAppRequest r) {
+    CommandResult createMobileApp(CreateMobileAppRequest r) {
         auto existing = repo.findByBundleId(r.bundleId);
         if (!existing.isNull)
             return CommandResult(false, "", "App with this bundle ID already exists");
@@ -44,7 +44,7 @@ class ManageMobileAppsUseCase { // TODO: UIMUseCase {
         return CommandResult(true, app.id.value, "");
     }
 
-    CommandResult update(MobileAppId id, UpdateMobileAppRequest r) {
+    CommandResult updateMobileApp(MobileAppId id, UpdateMobileAppRequest r) {
         auto app = repo.findById(tenantId, id);
         if (app.isNull)
             return CommandResult(false, "", "App not found");
@@ -60,19 +60,23 @@ class ManageMobileAppsUseCase { // TODO: UIMUseCase {
         return CommandResult(true, app.id.value, "");
     }
 
-    MobileApp get_(MobileAppId id) {
+    MobileApp getMobileApp(MobileAppId id) {
         return repo.findById(tenantId, id);
     }
 
-    MobileApp[] listByTenant(TenantId tenantId) {
+    MobileApp[] listMobileAppsByTenant(TenantId tenantId) {
         return repo.findByTenant(tenantId);
     }
 
-    CommandResult remove(MobileAppId id) {
+    CommandResult deleteMobileApp(MobileAppId id) {
+        auto app = repo.findById(tenantId, id);
+        if (app.isNull)
+            return CommandResult(false, "", "App not found");
         repo.removeById(id);
+        return CommandResult(true, app.id.value, "");
     }
 
-    size_t countByTenant(TenantId tenantId) {
+    size_t countMobileAppsByTenant(TenantId tenantId) {
         return repo.countByTenant(tenantId);
     }
